agent/pyproject.toml
[project]
name = "perfride-recommend-agent"
version = "0.1.0"
description = "PerfRide training recommendation agent powered by Google ADK"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.0.0",
    "google-cloud-storage>=2.18.0",
    "google-genai>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
